Express REST API with MongoDB and Mongoose Part 1
Developed a full-fledged REST API server with Express, MongoDB, and Mongoose. You updated the Express application to use Mongoose to connect to the MongoDB server, and you integrated the Mongoose module that exports the Campsite Model into your application. You used the Campsite Model to interact with the MongoDB server in response to HTTP requests intercepted by the campsiteRouter, and deal appropriately with each one.